Improve clients lib
