+++ client/client.go
@@ -4,1969 +4,594 @@
 package client
 
 import (
-	"crypto/tls"
+	"context"
+	"errors"
 	"fmt"
-	"io"
-	"net"
-	"runtime"
+	"strconv"
+	"strings"
 	"sync"
-	"time"
 
-	"github.com/absmach/fluxmq/mqtt/packets"
-	v3 "github.com/absmach/fluxmq/mqtt/packets/v3"
-	v5 "github.com/absmach/fluxmq/mqtt/packets/v5"
+	amqpclient "github.com/absmach/fluxmq/client/amqp"
+	mqttclient "github.com/absmach/fluxmq/client/mqtt"
+	"github.com/absmach/fluxmq/topics"
 )
 
-var timeZero = time.Time{}
-
-// Client is a thread-safe MQTT client.
+// Client provides a unified messaging API over MQTT or AMQP transports.
 type Client struct {
-	opts *Options
+	defaultProtocol Protocol
 
-	// State management
-	state *stateManager
+	mqtt *mqttclient.Client
+	amqp *amqpclient.Client
 
-	// Connection
-	conn   net.Conn
-	connMu sync.RWMutex
-
-	// Server capabilities (MQTT 5.0)
-	serverCaps   *ServerCapabilities
-	serverCapsMu sync.RWMutex
-
-	// Topic aliases (MQTT 5.0)
-	topicAliases *topicAliasManager
-
-	// Queue subscriptions
-	queueSubs     *queueSubscriptions
-	queueAckCache *queueAckCache
-	subscriptions *subscriptionRegistry
-
-	// Pending operations
-	pending *pendingStore
-
-	// Message store for QoS 1/2
-	store MessageStore
-
-	// QoS 2 incoming messages waiting for PUBREL
-	qos2Incoming   map[uint16]*Message
-	qos2IncomingMu sync.Mutex
-
-	// Lifecycle
-	stopCh   chan struct{}
-	doneCh   chan struct{}
-	reconnMu sync.Mutex
-
-	// Keep-alive
-	pingStop     chan struct{}
-	lastActivity time.Time
-	activityMu   sync.Mutex
-	pingMu       sync.Mutex
-	waitingPing  bool
-	lastPingSent time.Time
-
-	// Message dispatching
-	msgCh       chan *Message
-	msgOverflow chan *Message
-	msgStop     chan struct{}
-	dispatchWg  sync.WaitGroup
-
-	// Server index for round-robin
-	serverIdx int
+	mu    sync.RWMutex
+	subs  map[string]MessageHandler
+	qsubs map[string]MessageHandler
 }
 
-// New creates a new MQTT client with the given options.
-func New(opts *Options) (*Client, error) {
+// New creates a unified client configured with one or both transports.
+func New(cfg *Config) (*Client, error) {
+	if cfg == nil {
+		cfg = NewConfig()
+	}
+	if cfg.MQTT == nil && cfg.AMQP == nil {
+		return nil, ErrNoTransport
+	}
+
+	c := &Client{
+		defaultProtocol: cfg.DefaultProtocol,
+		subs:            make(map[string]MessageHandler),
+		qsubs:           make(map[string]MessageHandler),
+	}
+
+	if cfg.MQTT != nil {
+		mqttClient, err := c.newMQTTClient(cfg.MQTT)
+		if err != nil {
+			return nil, err
+		}
+		c.mqtt = mqttClient
+	}
+	if cfg.AMQP != nil {
+		amqpClient, err := amqpclient.New(cfg.AMQP)
+		if err != nil {
+			return nil, err
+		}
+		c.amqp = amqpClient
+	}
+
+	return c, nil
+}
+
+// NewMQTT creates a unified client backed by the MQTT client.
+func NewMQTT(opts *mqttclient.Options) (*Client, error) {
 	if opts == nil {
-		return nil, ErrNilOptions
+		opts = mqttclient.NewOptions()
 	}
-
-	if err := opts.Validate(); err != nil {
-		return nil, err
-	}
-
-	store := opts.Store
-	if store == nil {
-		store = NewMemoryStore()
-	}
-
-	return &Client{
-		opts:          opts,
-		state:         newStateManager(),
-		pending:       newPendingStore(opts.MaxInflight),
-		store:         store,
-		qos2Incoming:  make(map[uint16]*Message),
-		queueSubs:     newQueueSubscriptions(),
-		queueAckCache: newQueueAckCache(5 * time.Minute),
-		subscriptions: newSubscriptionRegistry(),
-	}, nil
+	return New(NewConfig().SetMQTT(opts).SetDefaultProtocol(ProtocolMQTT))
 }
+
+// NewAMQP creates a unified client backed by the AMQP client.
+func NewAMQP(opts *amqpclient.Options) (*Client, error) {
+	if opts == nil {
+		opts = amqpclient.NewOptions()
+	}
+	return New(NewConfig().SetAMQP(opts).SetDefaultProtocol(ProtocolAMQP))
+}
+
+// MQTT returns the underlying MQTT client, if configured.
+func (c *Client) MQTT() *mqttclient.Client { return c.mqtt }
+
+// AMQP returns the underlying AMQP client, if configured.
+func (c *Client) AMQP() *amqpclient.Client { return c.amqp }
 
 // Connect establishes a connection to the broker.
-func (c *Client) Connect() error {
-	if c.state.isClosed() {
-		return ErrClientClosed
+func (c *Client) Connect(ctx context.Context) error {
+	if c.mqtt == nil && c.amqp == nil {
+		return ErrNoTransport
 	}
+	return doWithContext(ctx, func() error {
+		if c.mqtt != nil {
+			if err := c.mqtt.Connect(); err != nil {
+				return fmt.Errorf("%w: %v", ErrConnectFailed, err)
+			}
+		}
+		if c.amqp != nil {
+			if err := c.amqp.Connect(); err != nil {
+				if c.mqtt != nil {
+					_ = c.mqtt.Close()
+				}
+				return fmt.Errorf("%w: %v", ErrConnectFailed, err)
+			}
+		}
+		return nil
+	})
+}
 
-	if !c.state.transitionFrom(StateConnecting, StateDisconnected, StateReconnecting) {
-		return ErrAlreadyConnected
+// Close terminates the connection.
+func (c *Client) Close(ctx context.Context) error {
+	if c.mqtt == nil && c.amqp == nil {
+		return ErrNoTransport
 	}
+	return doWithContext(ctx, func() error {
+		var errs []error
+		if c.mqtt != nil {
+			if err := c.mqtt.Close(); err != nil {
+				errs = append(errs, fmt.Errorf("%w: %v", ErrCloseFailed, err))
+			}
+		}
+		if c.amqp != nil {
+			if err := c.amqp.Close(); err != nil {
+				errs = append(errs, fmt.Errorf("%w: %v", ErrCloseFailed, err))
+			}
+		}
+		return errors.Join(errs...)
+	})
+}
 
-	err := c.doConnect()
+// IsConnected reports whether the client is connected.
+func (c *Client) IsConnected() bool {
+	switch {
+	case c.mqtt != nil && c.amqp != nil:
+		return c.mqtt.IsConnected() && c.amqp.IsConnected()
+	case c.mqtt != nil:
+		return c.mqtt.IsConnected()
+	case c.amqp != nil:
+		return c.amqp.IsConnected()
+	default:
+		return false
+	}
+}
+
+// Publish publishes a message to a topic.
+func (c *Client) Publish(ctx context.Context, topic string, payload []byte, opts ...Option) error {
+	if c.mqtt == nil && c.amqp == nil {
+		return ErrNoTransport
+	}
+	po := buildPublishOptions(opts)
+	protocol, err := c.resolveProtocol(po.Protocol)
 	if err != nil {
-		c.state.set(StateDisconnected)
 		return err
 	}
 
-	c.state.set(StateConnected)
-
-	// Start background goroutines
-	c.stopCh = make(chan struct{})
-	c.doneCh = make(chan struct{})
-	c.startDispatcher()
-	go c.readLoop()
-
-	if c.opts.KeepAlive > 0 {
-		c.startKeepAlive()
-	}
-
-	// Restore client-side state on reconnect (subscriptions and QoS in-flight).
-	c.restoreState()
-
-	// Callback
-	if c.opts.OnConnect != nil {
-		go c.opts.OnConnect()
-	}
-
-	return nil
-}
-
-func (c *Client) doConnect() error {
-	// Try each server in order
-	var lastErr error
-	for i := 0; i < len(c.opts.Servers); i++ {
-		idx := (c.serverIdx + i) % len(c.opts.Servers)
-		addr := c.opts.Servers[idx]
-
-		err := c.connectToServer(addr)
-		if err == nil {
-			c.serverIdx = idx
+	return doWithContext(ctx, func() error {
+		if protocol == ProtocolMQTT {
+			qos := byte(0)
+			if po.QoS != nil {
+				qos = *po.QoS
+			}
+			retain := false
+			if po.Retain != nil {
+				retain = *po.Retain
+			}
+			msg := mqttclient.NewMessage(topic, payload, qos, retain)
+			msg.UserProperties = mqttUserProps(po.Properties)
+			if err := c.mqtt.PublishMessageContext(ctx, msg); err != nil {
+				return fmt.Errorf("%w: %v", ErrPublishFailed, err)
+			}
 			return nil
 		}
-		lastErr = err
-	}
 
-	if lastErr != nil {
-		return fmt.Errorf("%w: %v", ErrConnectFailed, lastErr)
-	}
-	return ErrConnectFailed
+		amqpOpts := &amqpclient.PublishOptions{
+			Exchange:   po.Exchange,
+			RoutingKey: po.RoutingKey,
+			Topic:      topic,
+			Payload:    payload,
+			Properties: amqpHeaders(po.Properties),
+		}
+		if po.Mandatory != nil {
+			amqpOpts.Mandatory = *po.Mandatory
+		}
+		if po.Immediate != nil {
+			amqpOpts.Immediate = *po.Immediate
+		}
+		if err := c.amqp.PublishWithOptionsContext(ctx, amqpOpts); err != nil {
+			return fmt.Errorf("%w: %v", ErrPublishFailed, err)
+		}
+		return nil
+	})
 }
 
-func (c *Client) connectToServer(addr string) error {
-	// Establish TCP connection
-	var conn net.Conn
-	var err error
-
-	dialer := &net.Dialer{Timeout: c.opts.ConnectTimeout}
-	if c.opts.TLSConfig != nil {
-		conn, err = tls.DialWithDialer(dialer, "tcp", addr, c.opts.TLSConfig)
-	} else {
-		conn, err = dialer.Dial("tcp", addr)
+// Subscribe subscribes to a topic and routes matching messages to handler.
+func (c *Client) Subscribe(ctx context.Context, topic string, handler MessageHandler, opts ...Option) error {
+	if c.mqtt == nil && c.amqp == nil {
+		return ErrNoTransport
 	}
+	if handler == nil {
+		return fmt.Errorf("%w: handler cannot be nil", ErrSubscribeFailed)
+	}
+
+	so := buildSubscribeOptions(opts)
+	protocol, err := c.resolveProtocol(so.Protocol)
+	if err != nil {
+		return err
+	}
+	return doWithContext(ctx, func() error {
+		if protocol == ProtocolMQTT {
+			qos := byte(0)
+			if so.QoS != nil {
+				qos = *so.QoS
+			}
+			if err := c.mqtt.SubscribeSingle(topic, qos); err != nil {
+				return fmt.Errorf("%w: %v", ErrSubscribeFailed, err)
+			}
+			c.mu.Lock()
+			c.subs[topic] = handler
+			c.mu.Unlock()
+			return nil
+		}
+
+		autoAck := true
+		if so.AutoAck != nil {
+			autoAck = *so.AutoAck
+		}
+		subOpts := &amqpclient.SubscribeOptions{Topic: topic, AutoAck: autoAck}
+		if err := c.amqp.SubscribeWithOptions(subOpts, func(msg *amqpclient.Message) {
+			handler(amqpToMessage(msg, ""))
+		}); err != nil {
+			return fmt.Errorf("%w: %v", ErrSubscribeFailed, err)
+		}
+		return nil
+	})
+}
+
+// Unsubscribe removes a topic subscription.
+func (c *Client) Unsubscribe(ctx context.Context, topic string, opts ...Option) error {
+	if c.mqtt == nil && c.amqp == nil {
+		return ErrNoTransport
+	}
+	so := buildSubscribeOptions(opts)
+	protocol, err := c.resolveProtocol(so.Protocol)
 	if err != nil {
 		return err
 	}
 
-	// Send CONNECT packet
-	if err := c.sendConnect(conn); err != nil {
-		conn.Close()
+	return doWithContext(ctx, func() error {
+		if protocol == ProtocolMQTT {
+			if err := c.mqtt.Unsubscribe(topic); err != nil {
+				return fmt.Errorf("%w: %v", ErrUnsubFailed, err)
+			}
+			c.mu.Lock()
+			delete(c.subs, topic)
+			c.mu.Unlock()
+			return nil
+		}
+
+		if err := c.amqp.Unsubscribe(topic); err != nil {
+			return fmt.Errorf("%w: %v", ErrUnsubFailed, err)
+		}
+		return nil
+	})
+}
+
+// PublishToQueue publishes a message to a durable queue.
+func (c *Client) PublishToQueue(ctx context.Context, queue string, payload []byte, opts ...Option) error {
+	if c.mqtt == nil && c.amqp == nil {
+		return ErrNoTransport
+	}
+	po := buildPublishOptions(opts)
+	protocol, err := c.resolveProtocol(po.Protocol)
+	if err != nil {
 		return err
 	}
 
-	// Read CONNACK
-	code, err := c.readConnAck(conn)
+	return doWithContext(ctx, func() error {
+		if protocol == ProtocolMQTT {
+			qos := byte(1)
+			if po.QoS != nil {
+				qos = *po.QoS
+			}
+			qopts := &mqttclient.QueuePublishOptions{
+				QueueName:  queue,
+				Payload:    payload,
+				Properties: mqttUserProps(po.Properties),
+				QoS:        qos,
+			}
+			if err := c.mqtt.PublishToQueueWithOptionsContext(ctx, qopts); err != nil {
+				return fmt.Errorf("%w: %v", ErrQueuePublishFailed, err)
+			}
+			return nil
+		}
+
+		qopts := &amqpclient.QueuePublishOptions{
+			QueueName:  queue,
+			Payload:    payload,
+			Properties: amqpHeaders(po.Properties),
+		}
+		if err := c.amqp.PublishToQueueWithOptionsContext(ctx, qopts); err != nil {
+			return fmt.Errorf("%w: %v", ErrQueuePublishFailed, err)
+		}
+		return nil
+	})
+}
+
+// SubscribeToQueue subscribes to a queue with a consumer group.
+func (c *Client) SubscribeToQueue(ctx context.Context, queue, group string, handler MessageHandler, opts ...Option) error {
+	if c.mqtt == nil && c.amqp == nil {
+		return ErrNoTransport
+	}
+	if handler == nil {
+		return fmt.Errorf("%w: handler cannot be nil", ErrQueueSubscribeFailed)
+	}
+	so := buildSubscribeOptions(opts)
+	protocol, err := c.resolveProtocol(so.Protocol)
 	if err != nil {
-		conn.Close()
 		return err
 	}
-	if code != ConnAccepted {
-		conn.Close()
-		return code
-	}
 
-	c.connMu.Lock()
-	c.conn = conn
-	c.connMu.Unlock()
+	return doWithContext(ctx, func() error {
+		if protocol == ProtocolMQTT {
+			if err := c.mqtt.SubscribeToQueue(queue, group, func(msg *mqttclient.QueueMessage) {
+				handler(mqttQueueToMessage(msg, queue))
+			}); err != nil {
+				return fmt.Errorf("%w: %v", ErrQueueSubscribeFailed, err)
+			}
+			c.mu.Lock()
+			c.qsubs[queue] = handler
+			c.mu.Unlock()
+			return nil
+		}
 
-	c.updateActivity()
-	return nil
+		if err := c.amqp.SubscribeToQueue(queue, group, func(msg *amqpclient.QueueMessage) {
+			handler(amqpQueueToMessage(msg, queue))
+		}); err != nil {
+			return fmt.Errorf("%w: %v", ErrQueueSubscribeFailed, err)
+		}
+		return nil
+	})
 }
 
-func (c *Client) sendConnect(conn net.Conn) error {
-	conn.SetWriteDeadline(time.Now().Add(c.opts.WriteTimeout))
-	defer conn.SetWriteDeadline(time.Time{})
-
-	keepAlive := uint16(c.opts.KeepAlive.Seconds())
-
-	if c.opts.ProtocolVersion == 5 {
-		pkt := &v5.Connect{
-			FixedHeader:     packets.FixedHeader{PacketType: packets.ConnectType},
-			ClientID:        c.opts.ClientID,
-			KeepAlive:       keepAlive,
-			ProtocolName:    "MQTT",
-			ProtocolVersion: 5,
-			CleanStart:      c.opts.CleanSession,
-		}
-
-		if c.opts.Username != "" {
-			pkt.UsernameFlag = true
-			pkt.Username = c.opts.Username
-		}
-		if c.opts.Password != "" {
-			pkt.PasswordFlag = true
-			pkt.Password = []byte(c.opts.Password)
-		}
-
-		if c.opts.Will != nil {
-			pkt.WillFlag = true
-			pkt.WillQoS = c.opts.Will.QoS
-			pkt.WillRetain = c.opts.Will.Retain
-			pkt.WillTopic = c.opts.Will.Topic
-			pkt.WillPayload = c.opts.Will.Payload
-
-			// Set v5 Will properties
-			if c.opts.Will.WillDelayInterval > 0 || c.opts.Will.PayloadFormat != nil ||
-				c.opts.Will.MessageExpiry > 0 || c.opts.Will.ContentType != "" ||
-				c.opts.Will.ResponseTopic != "" || len(c.opts.Will.CorrelationData) > 0 ||
-				len(c.opts.Will.UserProperties) > 0 {
-				pkt.WillProperties = &v5.WillProperties{}
-
-				if c.opts.Will.WillDelayInterval > 0 {
-					pkt.WillProperties.WillDelayInterval = &c.opts.Will.WillDelayInterval
-				}
-
-				if c.opts.Will.PayloadFormat != nil {
-					pkt.WillProperties.PayloadFormat = c.opts.Will.PayloadFormat
-				}
-
-				if c.opts.Will.MessageExpiry > 0 {
-					pkt.WillProperties.MessageExpiry = &c.opts.Will.MessageExpiry
-				}
-
-				if c.opts.Will.ContentType != "" {
-					pkt.WillProperties.ContentType = c.opts.Will.ContentType
-				}
-
-				if c.opts.Will.ResponseTopic != "" {
-					pkt.WillProperties.ResponseTopic = c.opts.Will.ResponseTopic
-				}
-
-				if len(c.opts.Will.CorrelationData) > 0 {
-					pkt.WillProperties.CorrelationData = c.opts.Will.CorrelationData
-				}
-
-				if len(c.opts.Will.UserProperties) > 0 {
-					pkt.WillProperties.User = make([]v5.User, 0, len(c.opts.Will.UserProperties))
-					for k, v := range c.opts.Will.UserProperties {
-						pkt.WillProperties.User = append(pkt.WillProperties.User, v5.User{Key: k, Value: v})
-					}
-				}
-			}
-		}
-
-		// Set v5 Connect properties
-		if c.opts.SessionExpiry > 0 || c.opts.ReceiveMaximum > 0 ||
-			c.opts.MaximumPacketSize > 0 || c.opts.TopicAliasMaximum > 0 ||
-			c.opts.RequestResponseInfo || !c.opts.RequestProblemInfo ||
-			c.opts.AuthMethod != "" {
-			if pkt.Properties == nil {
-				pkt.Properties = &v5.ConnectProperties{}
-			}
-
-			if c.opts.SessionExpiry > 0 {
-				pkt.Properties.SessionExpiryInterval = &c.opts.SessionExpiry
-			}
-
-			if c.opts.ReceiveMaximum > 0 {
-				pkt.Properties.ReceiveMaximum = &c.opts.ReceiveMaximum
-			}
-
-			if c.opts.MaximumPacketSize > 0 {
-				pkt.Properties.MaximumPacketSize = &c.opts.MaximumPacketSize
-			}
-
-			if c.opts.TopicAliasMaximum > 0 {
-				pkt.Properties.TopicAliasMaximum = &c.opts.TopicAliasMaximum
-			}
-
-			if c.opts.RequestResponseInfo {
-				one := byte(1)
-				pkt.Properties.RequestResponseInfo = &one
-			}
-
-			// RequestProblemInfo defaults to true, only set if explicitly false
-			if !c.opts.RequestProblemInfo {
-				zero := byte(0)
-				pkt.Properties.RequestProblemInfo = &zero
-			}
-
-			if c.opts.AuthMethod != "" {
-				pkt.Properties.AuthMethod = c.opts.AuthMethod
-				pkt.Properties.AuthData = c.opts.AuthData
-			}
-		}
-
-		c.updateActivity()
-		return pkt.Pack(conn)
+// UnsubscribeFromQueue removes a queue subscription.
+func (c *Client) UnsubscribeFromQueue(ctx context.Context, queue string, opts ...Option) error {
+	if c.mqtt == nil && c.amqp == nil {
+		return ErrNoTransport
 	}
-
-	// MQTT 3.1.1
-	pkt := &v3.Connect{
-		FixedHeader:     packets.FixedHeader{PacketType: packets.ConnectType},
-		ClientID:        c.opts.ClientID,
-		KeepAlive:       keepAlive,
-		ProtocolName:    "MQTT",
-		ProtocolVersion: 4,
-		CleanSession:    c.opts.CleanSession,
-	}
-
-	if c.opts.Username != "" {
-		pkt.UsernameFlag = true
-		pkt.Username = c.opts.Username
-	}
-	if c.opts.Password != "" {
-		pkt.PasswordFlag = true
-		pkt.Password = []byte(c.opts.Password)
-	}
-
-	if c.opts.Will != nil {
-		pkt.WillFlag = true
-		pkt.WillQoS = c.opts.Will.QoS
-		pkt.WillRetain = c.opts.Will.Retain
-		pkt.WillTopic = c.opts.Will.Topic
-		pkt.WillMessage = c.opts.Will.Payload
-	}
-
-	return pkt.Pack(conn)
-}
-
-func (c *Client) readConnAck(conn net.Conn) (ConnAckCode, error) {
-	conn.SetReadDeadline(time.Now().Add(c.opts.ConnectTimeout))
-	defer conn.SetReadDeadline(time.Time{})
-
-	if c.opts.ProtocolVersion == 5 {
-		// Enhanced auth may require multiple AUTH packet exchanges
-		// before the final CONNACK arrives.
-		for {
-			pkt, _, _, err := v5.ReadPacket(conn)
-			if err != nil {
-				return 0, err
-			}
-
-			if auth, ok := pkt.(*v5.Auth); ok {
-				if err := c.handleConnectAuth(conn, auth); err != nil {
-					return 0, err
-				}
-				continue
-			}
-
-			ack, ok := pkt.(*v5.ConnAck)
-			if !ok {
-				return 0, ErrUnexpectedPacket
-			}
-
-			// Parse and store server capabilities
-			caps := parseConnAckProperties(ack.Properties)
-			c.serverCapsMu.Lock()
-			c.serverCaps = caps
-			c.serverCapsMu.Unlock()
-
-			// Initialize topic alias manager with server's limits
-			c.topicAliases = newTopicAliasManager(
-				c.opts.TopicAliasMaximum, // client accepts from server
-				caps.TopicAliasMaximum,   // server accepts from client
-			)
-
-			// Invoke callback if set
-			if c.opts.OnServerCapabilities != nil {
-				c.opts.OnServerCapabilities(caps)
-			}
-
-			return ConnAckCode(ack.ReasonCode), nil
-		}
-	}
-
-	pkt, err := v3.ReadPacket(conn)
+	so := buildSubscribeOptions(opts)
+	protocol, err := c.resolveProtocol(so.Protocol)
 	if err != nil {
-		return 0, err
+		return err
 	}
-	ack, ok := pkt.(*v3.ConnAck)
-	if !ok {
-		return 0, ErrUnexpectedPacket
-	}
-	return ConnAckCode(ack.ReturnCode), nil
+
+	return doWithContext(ctx, func() error {
+		if protocol == ProtocolMQTT {
+			if err := c.mqtt.UnsubscribeFromQueue(queue); err != nil {
+				return fmt.Errorf("%w: %v", ErrQueueUnsubFailed, err)
+			}
+			c.mu.Lock()
+			delete(c.qsubs, queue)
+			c.mu.Unlock()
+			return nil
+		}
+
+		if err := c.amqp.UnsubscribeFromQueue(queue); err != nil {
+			return fmt.Errorf("%w: %v", ErrQueueUnsubFailed, err)
+		}
+		return nil
+	})
 }
 
-// Disconnect gracefully disconnects from the broker.
-func (c *Client) Disconnect() error {
-	return c.DisconnectWithReason(0, 0, "")
+func (c *Client) newMQTTClient(opts *mqttclient.Options) (*mqttclient.Client, error) {
+	if opts == nil {
+		opts = mqttclient.NewOptions()
+	}
+
+	userOnMessageV2 := opts.OnMessageV2
+	userOnMessage := opts.OnMessage
+
+	cloned := *opts
+	cloned.OnMessageV2 = func(msg *mqttclient.Message) {
+		c.dispatchMQTT(msg)
+		if userOnMessageV2 != nil {
+			userOnMessageV2(msg)
+			return
+		}
+		if userOnMessage != nil {
+			userOnMessage(msg.Topic, msg.Payload, msg.QoS)
+		}
+	}
+	cloned.OnMessage = nil
+
+	return mqttclient.New(&cloned)
 }
 
-// DisconnectWithReason disconnects with an optional reason code (MQTT 5.0).
-// For MQTT 3.1.1, reasonCode and sessionExpiry are ignored.
-// Parameters:
-//   - reasonCode: MQTT 5.0 disconnect reason (0 = normal disconnect)
-//   - sessionExpiry: Update session expiry interval (0 = use current value, ignored if 0)
-//   - reasonString: Human-readable reason (empty = no reason string)
-func (c *Client) DisconnectWithReason(reasonCode byte, sessionExpiry uint32, reasonString string) error {
-	if !c.state.transition(StateConnected, StateDisconnecting) {
+func (c *Client) resolveProtocol(requested Protocol) (Protocol, error) {
+	switch requested {
+	case "":
+		return c.defaultProtocolOrFallback()
+	case ProtocolMQTT:
+		if c.mqtt == nil {
+			return "", fmt.Errorf("%w: %s", ErrNoRouteProtocol, ProtocolMQTT)
+		}
+		return ProtocolMQTT, nil
+	case ProtocolAMQP:
+		if c.amqp == nil {
+			return "", fmt.Errorf("%w: %s", ErrNoRouteProtocol, ProtocolAMQP)
+		}
+		return ProtocolAMQP, nil
+	default:
+		return "", fmt.Errorf("%w: %q", ErrInvalidProtocol, requested)
+	}
+}
+
+func (c *Client) defaultProtocolOrFallback() (Protocol, error) {
+	if c.defaultProtocol == ProtocolMQTT && c.mqtt != nil {
+		return ProtocolMQTT, nil
+	}
+	if c.defaultProtocol == ProtocolAMQP && c.amqp != nil {
+		return ProtocolAMQP, nil
+	}
+	if c.mqtt != nil {
+		return ProtocolMQTT, nil
+	}
+	if c.amqp != nil {
+		return ProtocolAMQP, nil
+	}
+	return "", ErrNoTransport
+}
+
+func (c *Client) dispatchMQTT(msg *mqttclient.Message) {
+	if msg == nil {
+		return
+	}
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	for filter, handler := range c.subs {
+		if topics.TopicMatch(filter, msg.Topic) {
+			handler(mqttToMessage(msg))
+		}
+	}
+}
+
+func mqttToMessage(msg *mqttclient.Message) *Message {
+	if msg == nil {
 		return nil
 	}
 
-	c.stopKeepAlive()
-	c.sendDisconnectWithReason(reasonCode, sessionExpiry, reasonString)
-	c.cleanup(nil)
-	c.state.set(StateDisconnected)
-
-	return nil
-}
-
-func (c *Client) sendDisconnectWithReason(reasonCode byte, sessionExpiry uint32, reasonString string) {
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn == nil {
-		return
+	props := make(map[string]string)
+	props[propMQTTQoS] = strconv.Itoa(int(msg.QoS))
+	props[propMQTTRetain] = strconv.FormatBool(msg.Retain)
+	props[propMQTTDup] = strconv.FormatBool(msg.Dup)
+	for k, v := range msg.UserProperties {
+		props[propMQTTUserPref+k] = v
 	}
 
-	conn.SetWriteDeadline(time.Now().Add(c.opts.WriteTimeout))
-
-	if c.opts.ProtocolVersion == 5 {
-		pkt := &v5.Disconnect{
-			FixedHeader: packets.FixedHeader{PacketType: packets.DisconnectType},
-			ReasonCode:  reasonCode,
-		}
-
-		// Add properties if specified
-		if sessionExpiry > 0 || reasonString != "" {
-			pkt.Properties = &v5.DisconnectProperties{}
-
-			if sessionExpiry > 0 {
-				pkt.Properties.SessionExpiryInterval = &sessionExpiry
-			}
-
-			if reasonString != "" {
-				pkt.Properties.ReasonString = reasonString
-			}
-		}
-
-		pkt.Pack(conn)
-	} else {
-		pkt := &v3.Disconnect{
-			FixedHeader: packets.FixedHeader{PacketType: packets.DisconnectType},
-		}
-		pkt.Pack(conn)
+	return &Message{
+		Topic:      msg.Topic,
+		Payload:    msg.Payload,
+		Properties: props,
+		Timestamp:  msg.Timestamp,
 	}
 }
 
-// Close permanently closes the client.
-func (c *Client) Close() error {
-	c.state.set(StateClosed)
-	c.stopKeepAlive()
-	c.cleanup(ErrClientClosed)
-	if c.store != nil {
-		c.store.Close()
-	}
-	return nil
-}
-
-func (c *Client) cleanup(err error) {
-	// Stop read loop
-	if c.stopCh != nil {
-		close(c.stopCh)
-		<-c.doneCh
-		c.stopCh = nil
-		c.doneCh = nil
-	}
-
-	c.stopDispatcher()
-
-	// Close connection
-	c.connMu.Lock()
-	if c.conn != nil {
-		c.conn.Close()
-		c.conn = nil
-	}
-	c.connMu.Unlock()
-
-	// Clear pending operations
-	c.pending.clear(err)
-
-	// Clear QoS 2 state
-	c.qos2IncomingMu.Lock()
-	c.qos2Incoming = make(map[uint16]*Message)
-	c.qos2IncomingMu.Unlock()
-
-	// Reset topic aliases
-	if c.topicAliases != nil {
-		c.topicAliases.reset()
-	}
-
-	c.pingMu.Lock()
-	c.waitingPing = false
-	c.lastPingSent = time.Time{}
-	c.pingMu.Unlock()
-}
-
-// IsConnected returns true if the client is connected.
-func (c *Client) IsConnected() bool {
-	return c.state.isConnected()
-}
-
-// State returns the current client state.
-func (c *Client) State() State {
-	return c.state.get()
-}
-
-// Publish sends a message to the broker.
-func (c *Client) Publish(topic string, payload []byte, qos byte, retain bool) error {
-	if !c.state.isConnected() {
-		return ErrNotConnected
-	}
-	if qos > 2 {
-		return ErrInvalidQoS
-	}
-	if topic == "" {
-		return ErrInvalidTopic
-	}
-
-	msg := NewMessage(topic, payload, qos, retain)
-
-	if qos == 0 {
-		return c.sendPublish(msg, 0)
-	}
-
-	// QoS 1 or 2: need packet ID and wait for ack
-	packetID := c.pending.nextPacketID()
-	if packetID == 0 {
-		return ErrMaxInflight
-	}
-	msg.PacketID = packetID
-
-	// Store message for potential retransmission
-	if err := c.store.StoreOutbound(packetID, msg); err != nil {
-		return err
-	}
-
-	op, err := c.pending.add(packetID, pendingPublish, msg)
-	if err != nil {
-		c.store.DeleteOutbound(packetID)
-		return err
-	}
-
-	if err := c.sendPublish(msg, packetID); err != nil {
-		c.pending.remove(packetID)
-		c.store.DeleteOutbound(packetID)
-		return err
-	}
-
-	if err := op.wait(c.opts.AckTimeout); err != nil {
-		c.pending.remove(packetID)
-		c.store.DeleteOutbound(packetID)
-		return err
-	}
-	return nil
-}
-
-// PublishMessage sends a message with optional MQTT 5.0 publish properties.
-// For MQTT 3.1.1, publish properties are ignored.
-func (c *Client) PublishMessage(msg *Message) error {
+func mqttQueueToMessage(msg *mqttclient.QueueMessage, queue string) *Message {
 	if msg == nil {
-		return ErrInvalidMessage
-	}
-	if !c.state.isConnected() {
-		return ErrNotConnected
-	}
-	if msg.QoS > 2 {
-		return ErrInvalidQoS
-	}
-	if msg.Topic == "" {
-		return ErrInvalidTopic
+		return nil
 	}
 
-	if msg.QoS == 0 {
-		return c.sendPublish(msg, 0)
+	m := mqttToMessage(msg.Message)
+	if m == nil {
+		return nil
 	}
-
-	packetID := c.pending.nextPacketID()
-	if packetID == 0 {
-		return ErrMaxInflight
-	}
-
-	publishMsg := msg.Copy()
-	publishMsg.PacketID = packetID
-
-	if err := c.store.StoreOutbound(packetID, publishMsg); err != nil {
-		return err
-	}
-
-	op, err := c.pending.add(packetID, pendingPublish, publishMsg)
-	if err != nil {
-		c.store.DeleteOutbound(packetID)
-		return err
-	}
-
-	if err := c.sendPublish(publishMsg, packetID); err != nil {
-		c.pending.remove(packetID)
-		c.store.DeleteOutbound(packetID)
-		return err
-	}
-
-	if err := op.wait(c.opts.AckTimeout); err != nil {
-		c.pending.remove(packetID)
-		c.store.DeleteOutbound(packetID)
-		return err
-	}
-	return nil
+	m.Queue = queue
+	m.Offset = msg.Offset
+	m.ackFn = msg.Ack
+	m.nackFn = msg.Nack
+	m.rejectFn = msg.Reject
+	return m
 }
 
-// PublishAsync publishes in a background goroutine and returns a completion token.
-func (c *Client) PublishAsync(topic string, payload []byte, qos byte, retain bool) *PublishToken {
-	tok := &PublishToken{token: newToken()}
-	go func() {
-		tok.complete(c.Publish(topic, payload, qos, retain))
-	}()
-	return tok
+func amqpToMessage(msg *amqpclient.Message, queue string) *Message {
+	if msg == nil {
+		return nil
+	}
+
+	props := make(map[string]string)
+	for k, v := range msg.Headers {
+		props[propAMQPHeadersPref+k] = fmt.Sprint(v)
+	}
+
+	return &Message{
+		Topic:      msg.Topic,
+		Payload:    msg.Body,
+		Properties: props,
+		Timestamp:  msg.Timestamp,
+		Queue:      queue,
+		ackFn:      msg.Ack,
+		nackFn:     msg.Nack,
+		rejectFn:   msg.Reject,
+	}
 }
 
-// PublishMessageAsync publishes a message with properties in a background goroutine.
-func (c *Client) PublishMessageAsync(msg *Message) *PublishToken {
-	tok := &PublishToken{token: newToken()}
-	go func() {
-		tok.complete(c.PublishMessage(msg))
-	}()
-	return tok
+func amqpQueueToMessage(msg *amqpclient.QueueMessage, queue string) *Message {
+	if msg == nil {
+		return nil
+	}
+
+	props := make(map[string]string)
+	for k, v := range msg.Headers {
+		props[propAMQPHeadersPref+k] = fmt.Sprint(v)
+	}
+
+	m := &Message{
+		Topic:      msg.RoutingKey,
+		Payload:    msg.Body,
+		Properties: props,
+		Timestamp:  msg.Timestamp,
+		Queue:      queue,
+		ackFn:      msg.Ack,
+		nackFn:     msg.Nack,
+		rejectFn:   msg.Reject,
+	}
+
+	if offset, ok := msg.StreamOffset(); ok {
+		m.Offset = offset
+	}
+	return m
 }
 
-func (c *Client) sendPublish(msg *Message, packetID uint16) error {
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn == nil {
-		return ErrNotConnected
+func mqttUserProps(props map[string]string) map[string]string {
+	if len(props) == 0 {
+		return nil
 	}
 
-	conn.SetWriteDeadline(time.Now().Add(c.opts.WriteTimeout))
-	defer conn.SetWriteDeadline(time.Time{})
-
-	if c.opts.ProtocolVersion == 5 {
-		pkt := &v5.Publish{
-			FixedHeader: packets.FixedHeader{
-				PacketType: packets.PublishType,
-				QoS:        msg.QoS,
-				Retain:     msg.Retain,
-				Dup:        msg.Dup,
-			},
-			TopicName: msg.Topic,
-			Payload:   msg.Payload,
-			ID:        packetID,
-		}
-
-		if msg.PayloadFormat != nil || msg.MessageExpiry != nil || msg.ContentType != "" || msg.ResponseTopic != "" || len(msg.CorrelationData) > 0 || len(msg.UserProperties) > 0 {
-			if pkt.Properties == nil {
-				pkt.Properties = &v5.PublishProperties{}
-			}
-		}
-
-		if msg.PayloadFormat != nil {
-			pkt.Properties.PayloadFormat = msg.PayloadFormat
-		}
-		if msg.MessageExpiry != nil {
-			pkt.Properties.MessageExpiry = msg.MessageExpiry
-		}
-		if msg.ContentType != "" {
-			pkt.Properties.ContentType = msg.ContentType
-		}
-		if msg.ResponseTopic != "" {
-			pkt.Properties.ResponseTopic = msg.ResponseTopic
-		}
-		if len(msg.CorrelationData) > 0 {
-			pkt.Properties.CorrelationData = msg.CorrelationData
-		}
-
-		if len(msg.UserProperties) > 0 {
-			if pkt.Properties == nil {
-				pkt.Properties = &v5.PublishProperties{}
-			}
-			pkt.Properties.User = make([]v5.User, 0, len(msg.UserProperties))
-			for k, v := range msg.UserProperties {
-				pkt.Properties.User = append(pkt.Properties.User, v5.User{Key: k, Value: v})
-			}
-		}
-
-		// Apply topic alias if available
-		if c.topicAliases != nil {
-			if alias, isNew, ok := c.topicAliases.getOrAssignOutbound(msg.Topic); ok {
-				if pkt.Properties == nil {
-					pkt.Properties = &v5.PublishProperties{}
-				}
-				pkt.Properties.TopicAlias = &alias
-
-				// If using existing alias, clear topic name to save bandwidth
-				if !isNew {
-					pkt.TopicName = ""
-				}
-				// If new alias, send both topic and alias for registration
-			}
-		}
-
-		c.updateActivity()
-		return pkt.Pack(conn)
-	}
-
-	pkt := &v3.Publish{
-		FixedHeader: packets.FixedHeader{
-			PacketType: packets.PublishType,
-			QoS:        msg.QoS,
-			Retain:     msg.Retain,
-			Dup:        msg.Dup,
-		},
-		TopicName: msg.Topic,
-		Payload:   msg.Payload,
-		ID:        packetID,
-	}
-	c.updateActivity()
-	return pkt.Pack(conn)
-}
-
-// Subscribe subscribes to one or more topics.
-func (c *Client) Subscribe(topics map[string]byte) error {
-	return c.subscribe(topics, true)
-}
-
-func (c *Client) subscribe(topics map[string]byte, remember bool) error {
-	if !c.state.isConnected() {
-		return ErrNotConnected
-	}
-	if len(topics) == 0 {
-		return ErrInvalidTopic
-	}
-	for topic, qos := range topics {
-		if topic == "" {
-			return ErrInvalidTopic
-		}
-		if qos > 2 {
-			return ErrInvalidQoS
-		}
-	}
-
-	packetID := c.pending.nextPacketID()
-	if packetID == 0 {
-		return ErrMaxInflight
-	}
-
-	op, err := c.pending.add(packetID, pendingSubscribe, nil)
-	if err != nil {
-		return err
-	}
-
-	if err := c.sendSubscribe(packetID, topics); err != nil {
-		c.pending.remove(packetID)
-		return err
-	}
-
-	if err := op.wait(c.opts.AckTimeout); err != nil {
-		c.pending.remove(packetID)
-		return err
-	}
-
-	if remember {
-		for topic, qos := range topics {
-			c.subscriptions.setBasic(topic, qos)
-		}
-	}
-
-	return nil
-}
-
-// SubscribeSingle is a convenience method for subscribing to a single topic.
-func (c *Client) SubscribeSingle(topic string, qos byte) error {
-	return c.Subscribe(map[string]byte{topic: qos})
-}
-
-// SubscribeWithOptions subscribes with advanced MQTT 5.0 options.
-// For MQTT 3.1.1 connections, advanced options are ignored.
-func (c *Client) SubscribeWithOptions(opts ...*SubscribeOption) error {
-	return c.subscribeWithOptions(opts, true)
-}
-
-func (c *Client) subscribeWithOptions(opts []*SubscribeOption, remember bool) error {
-	if !c.state.isConnected() {
-		return ErrNotConnected
-	}
-	if len(opts) == 0 {
-		return ErrInvalidTopic
-	}
-	for _, opt := range opts {
-		if opt == nil {
-			return ErrInvalidSubscribeOpt
-		}
-		if opt.Topic == "" {
-			return ErrInvalidTopic
-		}
-		if opt.QoS > 2 {
-			return ErrInvalidQoS
-		}
-		if opt.RetainHandling > 2 {
-			return ErrInvalidSubscribeOpt
-		}
-	}
-
-	packetID := c.pending.nextPacketID()
-	if packetID == 0 {
-		return ErrMaxInflight
-	}
-
-	op, err := c.pending.add(packetID, pendingSubscribe, nil)
-	if err != nil {
-		return err
-	}
-
-	if err := c.sendSubscribeWithOptions(packetID, opts); err != nil {
-		c.pending.remove(packetID)
-		return err
-	}
-
-	if err := op.wait(c.opts.AckTimeout); err != nil {
-		c.pending.remove(packetID)
-		return err
-	}
-
-	if remember {
-		for _, opt := range opts {
-			c.subscriptions.setOption(opt)
-		}
-	}
-
-	return nil
-}
-
-func (c *Client) sendSubscribe(packetID uint16, topics map[string]byte) error {
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn == nil {
-		return ErrNotConnected
-	}
-
-	conn.SetWriteDeadline(time.Now().Add(c.opts.WriteTimeout))
-	defer conn.SetWriteDeadline(time.Time{})
-
-	if c.opts.ProtocolVersion == 5 {
-		opts := make([]v5.SubOption, 0, len(topics))
-		for topic, qos := range topics {
-			opts = append(opts, v5.SubOption{Topic: topic, MaxQoS: qos})
-		}
-		pkt := &v5.Subscribe{
-			FixedHeader: packets.FixedHeader{PacketType: packets.SubscribeType, QoS: 1},
-			ID:          packetID,
-			Opts:        opts,
-		}
-		c.updateActivity()
-		return pkt.Pack(conn)
-	}
-
-	ts := make([]v3.Topic, 0, len(topics))
-	for topic, qos := range topics {
-		ts = append(ts, v3.Topic{Name: topic, QoS: qos})
-	}
-	pkt := &v3.Subscribe{
-		FixedHeader: packets.FixedHeader{PacketType: packets.SubscribeType, QoS: 1},
-		ID:          packetID,
-		Topics:      ts,
-	}
-	c.updateActivity()
-	return pkt.Pack(conn)
-}
-
-func (c *Client) sendSubscribeWithOptions(packetID uint16, opts []*SubscribeOption) error {
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn == nil {
-		return ErrNotConnected
-	}
-
-	conn.SetWriteDeadline(time.Now().Add(c.opts.WriteTimeout))
-	defer conn.SetWriteDeadline(time.Time{})
-
-	if c.opts.ProtocolVersion == 5 {
-		v5Opts := make([]v5.SubOption, len(opts))
-		for i, opt := range opts {
-			v5Opts[i] = v5.SubOption{
-				Topic:  opt.Topic,
-				MaxQoS: opt.QoS,
-			}
-
-			// Set advanced options if specified
-			if opt.NoLocal {
-				noLocal := true
-				v5Opts[i].NoLocal = &noLocal
-			}
-
-			if opt.RetainAsPublished {
-				rap := true
-				v5Opts[i].RetainAsPublished = &rap
-			}
-
-			if opt.RetainHandling > 0 {
-				v5Opts[i].RetainHandling = &opt.RetainHandling
-			}
-		}
-
-		pkt := &v5.Subscribe{
-			FixedHeader: packets.FixedHeader{PacketType: packets.SubscribeType, QoS: 1},
-			ID:          packetID,
-			Opts:        v5Opts,
-		}
-
-		// Use subscription identifier from first option (applies to all topics in this packet)
-		if len(opts) > 0 && opts[0].SubscriptionID > 0 {
-			subID := int(opts[0].SubscriptionID)
-			pkt.Properties = &v5.SubscribeProperties{
-				SubscriptionIdentifier: &subID,
-			}
-		}
-
-		c.updateActivity()
-		return pkt.Pack(conn)
-	}
-
-	// For MQTT 3.1.1, ignore advanced options
-	ts := make([]v3.Topic, len(opts))
-	for i, opt := range opts {
-		ts[i] = v3.Topic{Name: opt.Topic, QoS: opt.QoS}
-	}
-	pkt := &v3.Subscribe{
-		FixedHeader: packets.FixedHeader{PacketType: packets.SubscribeType, QoS: 1},
-		ID:          packetID,
-		Topics:      ts,
-	}
-	c.updateActivity()
-	return pkt.Pack(conn)
-}
-
-// Unsubscribe unsubscribes from one or more topics.
-func (c *Client) Unsubscribe(topics ...string) error {
-	if !c.state.isConnected() {
-		return ErrNotConnected
-	}
-	if len(topics) == 0 {
-		return ErrInvalidTopic
-	}
-
-	packetID := c.pending.nextPacketID()
-	if packetID == 0 {
-		return ErrMaxInflight
-	}
-
-	op, err := c.pending.add(packetID, pendingUnsubscribe, nil)
-	if err != nil {
-		return err
-	}
-
-	if err := c.sendUnsubscribe(packetID, topics); err != nil {
-		c.pending.remove(packetID)
-		return err
-	}
-
-	if err := op.wait(c.opts.AckTimeout); err != nil {
-		c.pending.remove(packetID)
-		return err
-	}
-
-	c.subscriptions.remove(topics...)
-
-	return nil
-}
-
-func (c *Client) sendUnsubscribe(packetID uint16, topics []string) error {
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn == nil {
-		return ErrNotConnected
-	}
-
-	conn.SetWriteDeadline(time.Now().Add(c.opts.WriteTimeout))
-	defer conn.SetWriteDeadline(time.Time{})
-
-	if c.opts.ProtocolVersion == 5 {
-		pkt := &v5.Unsubscribe{
-			FixedHeader: packets.FixedHeader{PacketType: packets.UnsubscribeType, QoS: 1},
-			ID:          packetID,
-			Topics:      topics,
-		}
-		c.updateActivity()
-		return pkt.Pack(conn)
-	}
-
-	pkt := &v3.Unsubscribe{
-		FixedHeader: packets.FixedHeader{PacketType: packets.UnsubscribeType, QoS: 1},
-		ID:          packetID,
-		Topics:      topics,
-	}
-	c.updateActivity()
-	return pkt.Pack(conn)
-}
-
-// SubscribeAsync subscribes in a background goroutine and returns a completion token.
-func (c *Client) SubscribeAsync(topics map[string]byte) *SubscribeToken {
-	tok := &SubscribeToken{token: newToken()}
-	go func() {
-		tok.complete(c.Subscribe(topics))
-	}()
-	return tok
-}
-
-// SubscribeWithOptionsAsync subscribes with MQTT 5 options in a background goroutine.
-func (c *Client) SubscribeWithOptionsAsync(opts ...*SubscribeOption) *SubscribeToken {
-	tok := &SubscribeToken{token: newToken()}
-	go func() {
-		tok.complete(c.SubscribeWithOptions(opts...))
-	}()
-	return tok
-}
-
-// UnsubscribeAsync unsubscribes in a background goroutine and returns a completion token.
-func (c *Client) UnsubscribeAsync(topics ...string) *UnsubscribeToken {
-	tok := &UnsubscribeToken{token: newToken()}
-	go func() {
-		tok.complete(c.Unsubscribe(topics...))
-	}()
-	return tok
-}
-
-// readLoop reads packets from the connection.
-func (c *Client) readLoop() {
-	defer close(c.doneCh)
-
-	for {
-		select {
-		case <-c.stopCh:
-			return
+	out := make(map[string]string)
+	for k, v := range props {
+		switch {
+		case strings.HasPrefix(k, propMQTTUserPref):
+			out[strings.TrimPrefix(k, propMQTTUserPref)] = v
+		case strings.HasPrefix(k, "mqtt."):
+			continue
+		case strings.HasPrefix(k, "amqp."):
+			continue
 		default:
+			out[k] = v
 		}
-
-		c.connMu.RLock()
-		conn := c.conn
-		c.connMu.RUnlock()
-
-		if conn == nil {
-			return
-		}
-
-		var pkt packets.ControlPacket
-		var err error
-
-		if c.opts.ProtocolVersion == 5 {
-			pkt, _, _, err = v5.ReadPacket(conn)
-		} else {
-			pkt, err = v3.ReadPacket(conn)
-		}
-
-		if err != nil {
-			if err == io.EOF || c.state.get() != StateConnected {
-				return
-			}
-			c.handleConnectionLost(err)
-			return
-		}
-
-		c.updateActivity()
-		c.handlePacket(pkt)
 	}
+	if len(out) == 0 {
+		return nil
+	}
+	return out
 }
 
-func (c *Client) handlePacket(pkt packets.ControlPacket) {
-	switch pkt.Type() {
-	case packets.PublishType:
-		c.handlePublish(pkt)
-	case packets.PubAckType:
-		c.handlePubAck(pkt)
-	case packets.PubRecType:
-		c.handlePubRec(pkt)
-	case packets.PubRelType:
-		c.handlePubRel(pkt)
-	case packets.PubCompType:
-		c.handlePubComp(pkt)
-	case packets.SubAckType:
-		c.handleSubAck(pkt)
-	case packets.UnsubAckType:
-		c.handleUnsubAck(pkt)
-	case packets.PingRespType:
-		c.pingMu.Lock()
-		c.waitingPing = false
-		c.lastPingSent = time.Time{}
-		c.pingMu.Unlock()
-	case packets.DisconnectType:
-		c.handleServerDisconnect(pkt)
-	case packets.AuthType:
-		c.handleAuth(pkt)
+func amqpHeaders(props map[string]string) map[string]string {
+	if len(props) == 0 {
+		return nil
 	}
+
+	out := make(map[string]string)
+	for k, v := range props {
+		switch {
+		case strings.HasPrefix(k, propAMQPHeadersPref):
+			out[strings.TrimPrefix(k, propAMQPHeadersPref)] = v
+		case strings.HasPrefix(k, "amqp."):
+			continue
+		case strings.HasPrefix(k, "mqtt."):
+			continue
+		default:
+			out[k] = v
+		}
+	}
+	if len(out) == 0 {
+		return nil
+	}
+	return out
 }
 
-func (c *Client) handlePublish(pkt packets.ControlPacket) {
-	var msg *Message
-
-	if c.opts.ProtocolVersion == 5 {
-		p := pkt.(*v5.Publish)
-
-		topic := p.TopicName
-
-		// Handle topic alias
-		if c.topicAliases != nil && p.Properties != nil && p.Properties.TopicAlias != nil {
-			alias := *p.Properties.TopicAlias
-
-			if topic != "" {
-				// Server sent both topic and alias - register the mapping
-				c.topicAliases.registerInbound(alias, topic)
-			} else {
-				// Server sent only alias - resolve it
-				var ok bool
-				topic, ok = c.topicAliases.resolveInbound(alias)
-				if !ok {
-					// Unknown alias - protocol error, ignore message
-					return
-				}
-			}
-		}
-
-		msg = &Message{
-			Topic:    topic,
-			Payload:  p.Payload,
-			QoS:      p.QoS,
-			Retain:   p.Retain,
-			Dup:      p.Dup,
-			PacketID: p.ID,
-		}
-
-		// Parse v5 properties
-		if p.Properties != nil {
-			msg.PayloadFormat = p.Properties.PayloadFormat
-			msg.MessageExpiry = p.Properties.MessageExpiry
-			msg.ContentType = p.Properties.ContentType
-			msg.ResponseTopic = p.Properties.ResponseTopic
-			msg.CorrelationData = p.Properties.CorrelationData
-
-			if len(p.Properties.User) > 0 {
-				msg.UserProperties = make(map[string]string, len(p.Properties.User))
-				for _, u := range p.Properties.User {
-					msg.UserProperties[u.Key] = u.Value
-				}
-			}
-
-			if p.Properties.SubscriptionID != nil {
-				msg.SubscriptionIDs = []uint32{uint32(*p.Properties.SubscriptionID)}
-			}
-		}
-	} else {
-		p := pkt.(*v3.Publish)
-		msg = &Message{
-			Topic:    p.TopicName,
-			Payload:  p.Payload,
-			QoS:      p.QoS,
-			Retain:   p.Retain,
-			Dup:      p.Dup,
-			PacketID: p.ID,
-		}
+func doWithContext(ctx context.Context, fn func() error) error {
+	if ctx == nil {
+		return fn()
+	}
+	if err := ctx.Err(); err != nil {
+		return err
 	}
 
-	switch msg.QoS {
-	case 0:
-		c.deliverMessage(msg)
-	case 1:
-		c.deliverMessage(msg)
-		c.sendPubAck(msg.PacketID)
-	case 2:
-		c.qos2IncomingMu.Lock()
-		c.qos2Incoming[msg.PacketID] = msg
-		c.qos2IncomingMu.Unlock()
-		_ = c.store.StoreInbound(msg.PacketID, msg)
-		c.sendPubRec(msg.PacketID)
-	}
-}
-
-func (c *Client) deliverMessage(msg *Message) {
-	if msg == nil {
-		return
-	}
-
-	if c.msgCh == nil {
-		c.handleDeliveredMessage(msg)
-		return
-	}
-
-	if c.opts.OrderMatters {
-		c.msgCh <- msg
-		return
-	}
+	errCh := make(chan error, 1)
+	go func() {
+		errCh <- fn()
+	}()
 
 	select {
-	case c.msgCh <- msg:
-	default:
-		// Primary channel full, try overflow channel
-		select {
-		case c.msgOverflow <- msg:
-		default:
-			// Both channels full — fall back to direct handling without blocking.
-			go c.handleDeliveredMessage(msg)
-		}
+	case err := <-errCh:
+		return err
+	case <-ctx.Done():
+		return ctx.Err()
 	}
-}
-
-func (c *Client) handleDeliveredMessage(msg *Message) {
-	// Check if this is a queue message
-	if isQueueTopic(msg.Topic) {
-		c.handleQueueMessage(msg)
-		return
-	}
-
-	// Call OnMessageV2 if set (provides full message context)
-	if c.opts.OnMessageV2 != nil {
-		c.opts.OnMessageV2(msg)
-		return
-	}
-
-	// Fall back to OnMessage for backward compatibility
-	if c.opts.OnMessage != nil {
-		c.opts.OnMessage(msg.Topic, msg.Payload, msg.QoS)
-	}
-}
-
-func (c *Client) startDispatcher() {
-	c.stopDispatcher()
-
-	size := c.opts.MessageChanSize
-	if size <= 0 {
-		size = DefaultMessageChanSize
-	}
-
-	c.msgCh = make(chan *Message, size)
-	c.msgOverflow = make(chan *Message, size)
-	c.msgStop = make(chan struct{})
-
-	workers := 1
-	if !c.opts.OrderMatters {
-		workers = runtime.GOMAXPROCS(0)
-		if workers < 2 {
-			workers = 2
-		}
-		if workers > 8 {
-			workers = 8
-		}
-	}
-
-	for i := 0; i < workers; i++ {
-		c.dispatchWg.Add(1)
-		go func() {
-			defer c.dispatchWg.Done()
-			for {
-				select {
-				case <-c.msgStop:
-					return
-				case msg, ok := <-c.msgCh:
-					if !ok {
-						return
-					}
-					c.handleDeliveredMessage(msg)
-				case msg, ok := <-c.msgOverflow:
-					if !ok {
-						return
-					}
-					c.handleDeliveredMessage(msg)
-				}
-			}
-		}()
-	}
-}
-
-func (c *Client) stopDispatcher() {
-	if c.msgStop != nil {
-		close(c.msgStop)
-		c.msgStop = nil
-	}
-	if c.msgCh != nil {
-		close(c.msgCh)
-		c.msgCh = nil
-	}
-	if c.msgOverflow != nil {
-		close(c.msgOverflow)
-		c.msgOverflow = nil
-	}
-	c.dispatchWg.Wait()
-}
-
-// handleQueueMessage processes a queue message and calls the appropriate handler.
-func (c *Client) handleQueueMessage(msg *Message) {
-	// Get queue subscription
-	sub, ok := c.queueSubs.get(msg.Topic)
-	if !ok || sub.handler == nil {
-		// No handler registered, fall through to default handlers
-		if c.opts.OnMessageV2 != nil {
-			c.opts.OnMessageV2(msg)
-		} else if c.opts.OnMessage != nil {
-			c.opts.OnMessage(msg.Topic, msg.Payload, msg.QoS)
-		}
-		return
-	}
-
-	// Extract queue message metadata from user properties
-	var messageID string
-	var groupID string
-	var offset uint64
-
-	if msg.UserProperties != nil {
-		if msgID, ok := msg.UserProperties["message-id"]; ok {
-			messageID = msgID
-		}
-		if gid, ok := msg.UserProperties["group-id"]; ok {
-			groupID = gid
-		}
-		if off, ok := msg.UserProperties["offset"]; ok {
-			fmt.Sscanf(off, "%d", &offset)
-		} else if seq, ok := msg.UserProperties["sequence"]; ok {
-			fmt.Sscanf(seq, "%d", &offset)
-		}
-	}
-
-	if c.queueAckCache != nil {
-		c.queueAckCache.set(messageID, groupID)
-	}
-
-	// Create queue message with ack/nack/reject methods
-	queueMsg := &QueueMessage{
-		Message:   msg,
-		MessageID: messageID,
-		GroupID:   groupID,
-		Offset:    offset,
-		Sequence:  offset,
-		client:    c,
-		queueName: sub.queueName,
-	}
-
-	// Call handler
-	sub.handler(queueMsg)
-}
-
-func (c *Client) handlePubAck(pkt packets.ControlPacket) {
-	var packetID uint16
-	if c.opts.ProtocolVersion == 5 {
-		packetID = pkt.(*v5.PubAck).ID
-	} else {
-		packetID = pkt.(*v3.PubAck).ID
-	}
-	c.store.DeleteOutbound(packetID)
-	c.pending.complete(packetID, nil, nil)
-}
-
-func (c *Client) handlePubRec(pkt packets.ControlPacket) {
-	var packetID uint16
-	if c.opts.ProtocolVersion == 5 {
-		packetID = pkt.(*v5.PubRec).ID
-	} else {
-		packetID = pkt.(*v3.PubRec).ID
-	}
-
-	c.pending.updateQoS2State(packetID, 1) // Now waiting for PUBCOMP
-	c.sendPubRel(packetID)
-}
-
-func (c *Client) handlePubRel(pkt packets.ControlPacket) {
-	var packetID uint16
-	if c.opts.ProtocolVersion == 5 {
-		packetID = pkt.(*v5.PubRel).ID
-	} else {
-		packetID = pkt.(*v3.PubRel).ID
-	}
-
-	c.qos2IncomingMu.Lock()
-	msg, exists := c.qos2Incoming[packetID]
-	if exists {
-		delete(c.qos2Incoming, packetID)
-	}
-	c.qos2IncomingMu.Unlock()
-
-	if !exists {
-		if storedMsg, ok := c.store.GetInbound(packetID); ok {
-			msg = storedMsg
-			exists = true
-		}
-	}
-
-	if exists {
-		_ = c.store.DeleteInbound(packetID)
-	}
-
-	if exists && msg != nil {
-		c.deliverMessage(msg)
-	}
-
-	c.sendPubComp(packetID)
-}
-
-func (c *Client) handlePubComp(pkt packets.ControlPacket) {
-	var packetID uint16
-	if c.opts.ProtocolVersion == 5 {
-		packetID = pkt.(*v5.PubComp).ID
-	} else {
-		packetID = pkt.(*v3.PubComp).ID
-	}
-	c.store.DeleteOutbound(packetID)
-	c.pending.complete(packetID, nil, nil)
-}
-
-func (c *Client) handleSubAck(pkt packets.ControlPacket) {
-	var packetID uint16
-	var returnCodes []byte
-
-	if c.opts.ProtocolVersion == 5 {
-		p := pkt.(*v5.SubAck)
-		packetID = p.ID
-		if p.ReasonCodes != nil {
-			returnCodes = *p.ReasonCodes
-		}
-	} else {
-		p := pkt.(*v3.SubAck)
-		packetID = p.ID
-		returnCodes = p.ReturnCodes
-	}
-
-	var err error
-	for _, rc := range returnCodes {
-		if c.opts.ProtocolVersion == 5 {
-			if rc >= 0x80 {
-				err = ErrSubscribeFailed
-				break
-			}
-			continue
-		}
-		if rc == 0x80 {
-			err = ErrSubscribeFailed
-			break
-		}
-	}
-
-	c.pending.complete(packetID, err, returnCodes)
-}
-
-func (c *Client) handleUnsubAck(pkt packets.ControlPacket) {
-	var packetID uint16
-	if c.opts.ProtocolVersion == 5 {
-		packetID = pkt.(*v5.UnsubAck).ID
-	} else {
-		packetID = pkt.(*v3.UnSubAck).ID
-	}
-	c.pending.complete(packetID, nil, nil)
-}
-
-func (c *Client) sendPubAck(packetID uint16) {
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn == nil {
-		return
-	}
-
-	if c.opts.ProtocolVersion == 5 {
-		pkt := &v5.PubAck{
-			FixedHeader: packets.FixedHeader{PacketType: packets.PubAckType},
-			ID:          packetID,
-		}
-		pkt.Pack(conn)
-	} else {
-		pkt := &v3.PubAck{
-			FixedHeader: packets.FixedHeader{PacketType: packets.PubAckType},
-			ID:          packetID,
-		}
-		pkt.Pack(conn)
-	}
-}
-
-func (c *Client) sendPubRec(packetID uint16) {
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn == nil {
-		return
-	}
-
-	if c.opts.ProtocolVersion == 5 {
-		pkt := &v5.PubRec{
-			FixedHeader: packets.FixedHeader{PacketType: packets.PubRecType},
-			ID:          packetID,
-		}
-		pkt.Pack(conn)
-	} else {
-		pkt := &v3.PubRec{
-			FixedHeader: packets.FixedHeader{PacketType: packets.PubRecType},
-			ID:          packetID,
-		}
-		pkt.Pack(conn)
-	}
-}
-
-func (c *Client) sendPubRel(packetID uint16) {
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn == nil {
-		return
-	}
-
-	if c.opts.ProtocolVersion == 5 {
-		pkt := &v5.PubRel{
-			FixedHeader: packets.FixedHeader{PacketType: packets.PubRelType, QoS: 1},
-			ID:          packetID,
-		}
-		pkt.Pack(conn)
-	} else {
-		pkt := &v3.PubRel{
-			FixedHeader: packets.FixedHeader{PacketType: packets.PubRelType, QoS: 1},
-			ID:          packetID,
-		}
-		pkt.Pack(conn)
-	}
-}
-
-func (c *Client) sendPubComp(packetID uint16) {
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn == nil {
-		return
-	}
-
-	if c.opts.ProtocolVersion == 5 {
-		pkt := &v5.PubComp{
-			FixedHeader: packets.FixedHeader{PacketType: packets.PubCompType},
-			ID:          packetID,
-		}
-		pkt.Pack(conn)
-	} else {
-		pkt := &v3.PubComp{
-			FixedHeader: packets.FixedHeader{PacketType: packets.PubCompType},
-			ID:          packetID,
-		}
-		pkt.Pack(conn)
-	}
-}
-
-func (c *Client) handleServerDisconnect(pkt packets.ControlPacket) {
-	if c.opts.ProtocolVersion != 5 {
-		return
-	}
-
-	d := pkt.(*v5.Disconnect)
-	reason := fmt.Sprintf("server disconnect: reason code 0x%02X", d.ReasonCode)
-	if d.Properties != nil && d.Properties.ReasonString != "" {
-		reason += " (" + d.Properties.ReasonString + ")"
-	}
-
-	c.handleConnectionLost(fmt.Errorf("%w: %s", ErrConnectionLost, reason))
-}
-
-// handleConnectAuth handles AUTH packets received during the CONNECT handshake.
-func (c *Client) handleConnectAuth(conn net.Conn, auth *v5.Auth) error {
-	if c.opts.OnAuth == nil {
-		return ErrNoAuthHandler
-	}
-
-	method := c.opts.AuthMethod
-	var data []byte
-	if auth.Properties != nil {
-		if auth.Properties.AuthMethod != "" {
-			if c.opts.AuthMethod != "" && c.opts.AuthMethod != auth.Properties.AuthMethod {
-				return ErrAuthMethodMismatch
-			}
-			method = auth.Properties.AuthMethod
-		}
-		data = auth.Properties.AuthData
-	}
-
-	if method == "" {
-		return ErrAuthMethodMissing
-	}
-
-	responseData, err := c.opts.OnAuth(auth.ReasonCode, method, data)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrAuthFailed, err)
-	}
-
-	resp := &v5.Auth{
-		FixedHeader: packets.FixedHeader{PacketType: packets.AuthType},
-		ReasonCode:  0x18, // Continue Authentication
-		Properties: &v5.AuthProperties{
-			AuthMethod: method,
-			AuthData:   responseData,
-		},
-	}
-
-	return resp.Pack(conn)
-}
-
-// handleAuth handles AUTH packets received during an active session (re-authentication).
-func (c *Client) handleAuth(pkt packets.ControlPacket) {
-	if c.opts.ProtocolVersion != 5 {
-		return
-	}
-
-	auth := pkt.(*v5.Auth)
-	if c.opts.OnAuth == nil {
-		return
-	}
-
-	method := c.opts.AuthMethod
-	var data []byte
-	if auth.Properties != nil {
-		if auth.Properties.AuthMethod != "" {
-			if c.opts.AuthMethod != "" && c.opts.AuthMethod != auth.Properties.AuthMethod {
-				c.handleConnectionLost(fmt.Errorf("%w: %s", ErrAuthMethodMismatch, auth.Properties.AuthMethod))
-				return
-			}
-			method = auth.Properties.AuthMethod
-		}
-		data = auth.Properties.AuthData
-	}
-
-	if method == "" {
-		c.handleConnectionLost(ErrAuthMethodMissing)
-		return
-	}
-
-	responseData, err := c.opts.OnAuth(auth.ReasonCode, method, data)
-	if err != nil {
-		c.handleConnectionLost(fmt.Errorf("%w: %v", ErrAuthFailed, err))
-		return
-	}
-
-	// Reason code 0x00 means success — no response needed
-	if auth.ReasonCode == 0x00 {
-		return
-	}
-
-	resp := &v5.Auth{
-		FixedHeader: packets.FixedHeader{PacketType: packets.AuthType},
-		ReasonCode:  0x18, // Continue Authentication
-		Properties: &v5.AuthProperties{
-			AuthMethod: method,
-			AuthData:   responseData,
-		},
-	}
-
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn != nil {
-		if err := resp.Pack(conn); err != nil {
-			c.handleConnectionLost(fmt.Errorf("%w: %v", ErrAuthFailed, err))
-		}
-	}
-}
-
-// SendAuth initiates re-authentication by sending an AUTH packet to the server (MQTT 5.0).
-func (c *Client) SendAuth(reasonCode byte, authData []byte) error {
-	if c.opts.ProtocolVersion != 5 {
-		return ErrAuthNotV5
-	}
-	if c.state.get() != StateConnected {
-		return ErrNotConnected
-	}
-	if c.opts.AuthMethod == "" {
-		return ErrAuthMethodMissing
-	}
-
-	pkt := &v5.Auth{
-		FixedHeader: packets.FixedHeader{PacketType: packets.AuthType},
-		ReasonCode:  reasonCode,
-		Properties: &v5.AuthProperties{
-			AuthMethod: c.opts.AuthMethod,
-			AuthData:   authData,
-		},
-	}
-
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn == nil {
-		return ErrNotConnected
-	}
-
-	return pkt.Pack(conn)
-}
-
-func (c *Client) handleConnectionLost(err error) {
-	if !c.state.transition(StateConnected, StateDisconnected) {
-		return
-	}
-
-	c.stopKeepAlive()
-	c.stopDispatcher()
-	c.pending.clear(ErrConnectionLost)
-
-	c.connMu.Lock()
-	if c.conn != nil {
-		c.conn.Close()
-		c.conn = nil
-	}
-	c.connMu.Unlock()
-
-	c.pingMu.Lock()
-	c.waitingPing = false
-	c.lastPingSent = time.Time{}
-	c.pingMu.Unlock()
-
-	if c.opts.OnConnectionLost != nil {
-		go c.opts.OnConnectionLost(err)
-	}
-
-	if c.opts.AutoReconnect && !c.state.isClosed() {
-		go c.reconnect()
-	}
-}
-
-func (c *Client) reconnect() {
-	c.reconnMu.Lock()
-	defer c.reconnMu.Unlock()
-
-	if !c.state.transition(StateDisconnected, StateReconnecting) {
-		return
-	}
-
-	delay := c.opts.ReconnectBackoff
-	attempt := 0
-
-	for !c.state.isClosed() {
-		attempt++
-
-		if c.opts.OnReconnecting != nil {
-			c.opts.OnReconnecting(attempt)
-		}
-
-		err := c.Connect()
-		if err == nil {
-			return
-		}
-
-		// Exponential backoff
-		time.Sleep(delay)
-		delay *= 2
-		if delay > c.opts.MaxReconnectWait {
-			delay = c.opts.MaxReconnectWait
-		}
-	}
-}
-
-func (c *Client) restoreState() {
-	if !c.state.isConnected() {
-		return
-	}
-
-	c.restoreSubscriptions()
-	c.restoreQueueSubscriptions()
-	c.restoreOutboundMessages()
-}
-
-func (c *Client) restoreSubscriptions() {
-	records := c.subscriptions.snapshot()
-	if len(records) == 0 {
-		return
-	}
-
-	for _, rec := range records {
-		if rec.opt != nil {
-			_ = c.subscribeWithOptions([]*SubscribeOption{rec.opt}, false)
-			continue
-		}
-		_ = c.subscribe(map[string]byte{rec.topic: rec.qos}, false)
-	}
-}
-
-func (c *Client) restoreQueueSubscriptions() {
-	c.queueSubs.mu.RLock()
-	subs := make([]*queueSubscription, 0, len(c.queueSubs.subs))
-	for _, sub := range c.queueSubs.subs {
-		copied := *sub
-		subs = append(subs, &copied)
-	}
-	c.queueSubs.mu.RUnlock()
-
-	for _, sub := range subs {
-		if c.opts.ProtocolVersion == 5 {
-			userProps := make(map[string]string)
-			if sub.consumerGroup != "" {
-				userProps["consumer-group"] = sub.consumerGroup
-			}
-			_ = c.subscribeWithUserProperties("$queue/"+sub.queueName, 1, userProps)
-			continue
-		}
-		_ = c.subscribe(map[string]byte{"$queue/" + sub.queueName: 1}, false)
-	}
-}
-
-func (c *Client) restoreOutboundMessages() {
-	msgs := c.store.GetAllOutbound()
-	if len(msgs) == 0 {
-		return
-	}
-
-	_ = c.store.Reset()
-
-	for _, msg := range msgs {
-		if msg == nil {
-			continue
-		}
-
-		if msg.QoS == 0 {
-			_ = c.sendPublish(msg, 0)
-			continue
-		}
-
-		packetID := c.pending.nextPacketID()
-		if packetID == 0 {
-			return
-		}
-
-		replay := msg.Copy()
-		replay.PacketID = packetID
-		replay.Dup = true
-
-		if err := c.store.StoreOutbound(packetID, replay); err != nil {
-			continue
-		}
-
-		if _, err := c.pending.add(packetID, pendingPublish, replay); err != nil {
-			_ = c.store.DeleteOutbound(packetID)
-			continue
-		}
-
-		if err := c.sendPublish(replay, packetID); err != nil {
-			c.pending.remove(packetID)
-			_ = c.store.DeleteOutbound(packetID)
-			continue
-		}
-	}
-}
-
-// Keep-alive management
-
-func (c *Client) startKeepAlive() {
-	stop := make(chan struct{})
-	c.activityMu.Lock()
-	c.pingStop = stop
-	c.activityMu.Unlock()
-
-	interval := c.opts.KeepAlive
-
-	go func() {
-		ticker := time.NewTicker(interval / 2)
-		defer ticker.Stop()
-
-		for {
-			select {
-			case <-stop:
-				return
-			case <-ticker.C:
-				c.activityMu.Lock()
-				idle := time.Since(c.lastActivity)
-				c.activityMu.Unlock()
-
-				if idle >= interval/2 {
-					c.sendPing()
-				}
-			}
-		}
-	}()
-}
-
-func (c *Client) stopKeepAlive() {
-	c.activityMu.Lock()
-	stop := c.pingStop
-	c.pingStop = nil
-	c.activityMu.Unlock()
-
-	if stop != nil {
-		close(stop)
-	}
-}
-
-func (c *Client) sendPing() {
-	c.connMu.RLock()
-	conn := c.conn
-	c.connMu.RUnlock()
-
-	if conn == nil {
-		return
-	}
-
-	timeout := c.opts.PingTimeout
-	if timeout <= 0 {
-		timeout = DefaultPingTimeout
-	}
-
-	c.pingMu.Lock()
-	if c.waitingPing {
-		sincePing := time.Since(c.lastPingSent)
-		c.pingMu.Unlock()
-		if sincePing >= timeout {
-			c.handleConnectionLost(ErrPingTimeout)
-		}
-		return
-	}
-	c.waitingPing = true
-	c.lastPingSent = time.Now().UTC()
-	c.pingMu.Unlock()
-
-	conn.SetWriteDeadline(time.Now().Add(c.opts.WriteTimeout))
-	defer conn.SetWriteDeadline(time.Time{})
-
-	if c.opts.ProtocolVersion == 5 {
-		pkt := &v5.PingReq{
-			FixedHeader: packets.FixedHeader{PacketType: packets.PingReqType},
-		}
-		pkt.Pack(conn)
-	} else {
-		pkt := &v3.PingReq{
-			FixedHeader: packets.FixedHeader{PacketType: packets.PingReqType},
-		}
-		pkt.Pack(conn)
-	}
-
-	c.updateActivity()
-}
-
-func (c *Client) updateActivity() {
-	c.activityMu.Lock()
-	defer c.activityMu.Unlock()
-	c.lastActivity = time.Now().UTC()
-}
-
-// ServerCapabilities returns the capabilities advertised by the server
-// in the CONNACK packet. This is only available for MQTT 5.0 connections.
-// Returns nil if not connected or using MQTT 3.1.1.
-func (c *Client) ServerCapabilities() *ServerCapabilities {
-	c.serverCapsMu.RLock()
-	defer c.serverCapsMu.RUnlock()
-	return c.serverCaps
 }
