+++ client/mqtt/subscribe.go
@@ -1,7 +1,7 @@
 // Copyright (c) Abstract Machines
 // SPDX-License-Identifier: Apache-2.0
 
-package client
+package mqtt
 
 // SubscribeOption represents subscription options for MQTT 5.0.
 type SubscribeOption struct {
