+++ tests/perf/loadgen/main.go
@@ -18,8 +18,9 @@ import (
 	"sync/atomic"
 	"time"
 
-	mqttclient "github.com/absmach/fluxmq/client"
+	msgclient "github.com/absmach/fluxmq/client"
 	amqpclient "github.com/absmach/fluxmq/client/amqp"
+	mqttclient "github.com/absmach/fluxmq/client/mqtt"
 	"github.com/absmach/fluxmq/topics"
 )
 
@@ -538,30 +539,32 @@ func runConfiguredTopicScenario(ctx context.Context, cfg runConfig, sc topicScen
 
 	switch sc.resolvedSubscriberProto {
 	case "mqtt":
-		subs, err := connectMQTTSubscribers(mqttSubscriberEndpoints, sc.Name, runID, sc.Subscribers, func(subID int, _ string, msg *mqttclient.Message) {
-			if !isCountablePayload(msg.Payload) {
-				return
-			}
-			msgID := extractMsgID(msg.Payload)
-			if seen.add(fmt.Sprintf("%d:%s", subID, msgID)) {
-				received.Add(1)
-			}
-		})
+		subs, err := connectMQTTSubscribers(mqttSubscriberEndpoints, sc.Name, runID, sc.Subscribers)
 		if err != nil {
 			return res, err
 		}
 		defer disconnectMQTTClients(subs)
 
 		for i, sub := range subs {
-			if err := sub.SubscribeSingle(filters[i], sc.resolvedQoS); err != nil {
+			filter := filters[i]
+			subID := i
+			if err := sub.Subscribe(ctx, filter, func(msg *msgclient.Message) {
+				if !isCountablePayload(msg.Payload) {
+					return
+				}
+				msgID := extractMsgID(msg.Payload)
+				if seen.add(fmt.Sprintf("%d:%s", subID, msgID)) {
+					received.Add(1)
+				}
+			}, msgclient.WithQoS(sc.resolvedQoS)); err != nil {
 				errCount.Add(1)
-				return res, fmt.Errorf("subscriber %d failed to subscribe %q: %w", i, filters[i], err)
+				return res, fmt.Errorf("subscriber %d failed to subscribe %q: %w", i, filter, err)
 			}
 		}
 	case "amqp":
-		subs, err := connectAMQPTopicSubscribersWithFilters(amqpSubscriberAddrs, sc.Name, runID, filters, func(subID int, msg *amqpclient.Message) {
-			if isCountablePayload(msg.Body) {
-				msgID := extractMsgID(msg.Body)
+		subs, err := connectAMQPTopicSubscribersWithFilters(ctx, amqpSubscriberAddrs, sc.Name, runID, filters, func(subID int, msg *msgclient.Message) {
+			if isCountablePayload(msg.Payload) {
+				msgID := extractMsgID(msg.Payload)
 				if seen.add(fmt.Sprintf("%d:%s", subID, msgID)) {
 					received.Add(1)
 				}
@@ -595,6 +598,10 @@ func runConfiguredTopicScenario(ctx context.Context, cfg runConfig, sc topicScen
 	}
 
 	var published int64
+	publishTimeout := cfg.DrainTimeout
+	if publishTimeout <= 0 {
+		publishTimeout = 30 * time.Second
+	}
 	switch sc.resolvedPublisherProto {
 	case "mqtt":
 		published = runMQTTPublishers(ctx, mqttPublishParams{
@@ -612,6 +619,7 @@ func runConfiguredTopicScenario(ctx context.Context, cfg runConfig, sc topicScen
 			ErrCount:             &errCount,
 			PublishedCounter:     &sent,
 			QoS:                  sc.resolvedQoS,
+			PublishTimeout:       publishTimeout,
 		})
 	case "amqp":
 		published = runAMQPTopicPublishers(ctx, amqpTopicPublishParams{
@@ -628,6 +636,7 @@ func runConfiguredTopicScenario(ctx context.Context, cfg runConfig, sc topicScen
 			OnPublishError:       onPublishError,
 			ErrCount:             &errCount,
 			PublishedCounter:     &sent,
+			PublishTimeout:       publishTimeout,
 		})
 	default:
 		return res, fmt.Errorf("unsupported publisher protocol %q", sc.resolvedPublisherProto)
@@ -737,18 +746,12 @@ func buildTopicMatchCounts(topicSet, filters []string) map[string]int64 {
 	return matchCounts
 }
 
-func connectMQTTSubscribers(endpoints []mqttEndpoint, scenario string, runID int64, count int, onMessage func(subID int, subName string, msg *mqttclient.Message)) ([]*mqttclient.Client, error) {
-	clients := make([]*mqttclient.Client, 0, count)
+func connectMQTTSubscribers(endpoints []mqttEndpoint, scenario string, runID int64, count int) ([]*msgclient.Client, error) {
+	clients := make([]*msgclient.Client, 0, count)
 	for i := 0; i < count; i++ {
 		endpoint := mqttEndpointByOffset(endpoints, i, subscriberNodeOffset)
 		clientID := fmt.Sprintf("%s-sub-%d-%d", scenario, runID, i)
-		sid := i
-		sname := clientID
-		client, err := connectMQTTClient(endpoint.Addr, endpoint.ProtocolVersion, clientID, func(msg *mqttclient.Message) {
-			if onMessage != nil {
-				onMessage(sid, sname, msg)
-			}
-		})
+		client, err := connectMQTTClient(endpoint.Addr, endpoint.ProtocolVersion, clientID)
 		if err != nil {
 			disconnectMQTTClients(clients)
 			return nil, fmt.Errorf("failed to connect MQTT subscriber %s to %s (v%d): %w", clientID, endpoint.Addr, endpoint.ProtocolVersion, err)
@@ -758,8 +761,8 @@ func connectMQTTSubscribers(endpoints []mqttEndpoint, scenario string, runID int
 	return clients, nil
 }
 
-func connectAMQPTopicSubscribersWithFilters(addrs []string, scenario string, runID int64, filters []string, handler func(subID int, msg *amqpclient.Message)) ([]*amqpclient.Client, error) {
-	clients := make([]*amqpclient.Client, 0, len(filters))
+func connectAMQPTopicSubscribersWithFilters(ctx context.Context, addrs []string, scenario string, runID int64, filters []string, handler func(subID int, msg *msgclient.Message)) ([]*msgclient.Client, error) {
+	clients := make([]*msgclient.Client, 0, len(filters))
 	for i, filter := range filters {
 		addr := addrByOffset(addrs, i, subscriberNodeOffset)
 		cid := i
@@ -768,11 +771,10 @@ func connectAMQPTopicSubscribersWithFilters(addrs []string, scenario string, run
 			closeAMQPClients(clients)
 			return nil, fmt.Errorf("failed to connect AMQP subscriber %d to %s: %w", i, addr, err)
 		}
-		subOpts := &amqpclient.SubscribeOptions{Topic: filter, AutoAck: false}
-		if err := client.SubscribeWithOptions(subOpts, func(msg *amqpclient.Message) {
+		if err := client.Subscribe(ctx, filter, func(msg *msgclient.Message) {
 			handler(cid, msg)
-		}); err != nil {
-			_ = client.Close()
+		}, msgclient.WithAutoAck(false)); err != nil {
+			_ = client.Close(context.Background())
 			closeAMQPClients(clients)
 			return nil, fmt.Errorf("failed AMQP subscribe for scenario=%s run=%d filter=%q: %w", scenario, runID, filter, err)
 		}
@@ -796,6 +798,7 @@ type mqttPublishParams struct {
 	ErrCount             *atomic.Int64
 	PublishedCounter     *atomic.Int64
 	QoS                  byte
+	PublishTimeout       time.Duration
 }
 
 func newPublishRNG(runID int64, publisherIdx int) *rand.Rand {
@@ -840,11 +843,11 @@ func sleepWithContext(ctx context.Context, d time.Duration) bool {
 }
 
 func runMQTTPublishers(ctx context.Context, params mqttPublishParams) int64 {
-	pubClients := make([]*mqttclient.Client, 0, params.Publishers)
+	pubClients := make([]*msgclient.Client, 0, params.Publishers)
 	for i := 0; i < params.Publishers; i++ {
 		endpoint := mqttEndpointByOffset(params.Endpoints, i, publisherNodeOffset)
 		clientID := fmt.Sprintf("%s-pub-%d-%d", params.Scenario, params.RunID, i)
-		client, err := connectMQTTClient(endpoint.Addr, endpoint.ProtocolVersion, clientID, nil)
+		client, err := connectMQTTClient(endpoint.Addr, endpoint.ProtocolVersion, clientID)
 		if err != nil {
 			if params.ErrCount != nil {
 				params.ErrCount.Add(1)
@@ -866,7 +869,7 @@ func runMQTTPublishers(ctx context.Context, params mqttPublishParams) int64 {
 	var wg sync.WaitGroup
 	for pubIdx, c := range pubClients {
 		wg.Add(1)
-		go func(idx int, client *mqttclient.Client) {
+		go func(idx int, client *msgclient.Client) {
 			defer wg.Done()
 			rng := newPublishRNG(params.RunID, idx)
 			if !sleepWithContext(ctx, initialPublishDelay(params.PublishJitter, rng)) {
@@ -885,7 +888,14 @@ func runMQTTPublishers(ctx context.Context, params mqttPublishParams) int64 {
 				}
 				msgID := fmt.Sprintf("%s-%d-%d", params.Scenario, idx, msgIdx)
 				payload := makePayload(msgID, params.PayloadSize)
-				if err := client.Publish(topic, payload, qos, false); err != nil {
+				pubCtx := ctx
+				cancel := func() {}
+				if params.PublishTimeout > 0 {
+					pubCtx, cancel = context.WithTimeout(ctx, params.PublishTimeout)
+				}
+				err := client.Publish(pubCtx, topic, payload, msgclient.WithQoS(qos))
+				cancel()
+				if err != nil {
 					if params.OnPublishError != nil {
 						params.OnPublishError(topic)
 					}
@@ -919,10 +929,11 @@ type amqpTopicPublishParams struct {
 	OnPublishError       func(topic string)
 	ErrCount             *atomic.Int64
 	PublishedCounter     *atomic.Int64
+	PublishTimeout       time.Duration
 }
 
 func runAMQPTopicPublishers(ctx context.Context, params amqpTopicPublishParams) int64 {
-	pubClients := make([]*amqpclient.Client, 0, params.Publishers)
+	pubClients := make([]*msgclient.Client, 0, params.Publishers)
 	for i := 0; i < params.Publishers; i++ {
 		addr := addrByOffset(params.Addrs, i, publisherNodeOffset)
 		client, err := connectAMQPClient(addr)
@@ -943,7 +954,7 @@ func runAMQPTopicPublishers(ctx context.Context, params amqpTopicPublishParams) 
 	var wg sync.WaitGroup
 	for pubIdx, c := range pubClients {
 		wg.Add(1)
-		go func(idx int, client *amqpclient.Client) {
+		go func(idx int, client *msgclient.Client) {
 			defer wg.Done()
 			rng := newPublishRNG(params.RunID, idx)
 			if !sleepWithContext(ctx, initialPublishDelay(params.PublishJitter, rng)) {
@@ -962,7 +973,14 @@ func runAMQPTopicPublishers(ctx context.Context, params amqpTopicPublishParams) 
 				}
 				msgID := fmt.Sprintf("%s-at-%d-%d", params.Scenario, idx, msgIdx)
 				payload := makePayload(msgID, params.PayloadSize)
-				if err := client.Publish(topic, payload); err != nil {
+				pubCtx := ctx
+				cancel := func() {}
+				if params.PublishTimeout > 0 {
+					pubCtx, cancel = context.WithTimeout(ctx, params.PublishTimeout)
+				}
+				err := client.Publish(pubCtx, topic, payload)
+				cancel()
+				if err != nil {
 					if params.OnPublishError != nil {
 						params.OnPublishError(topic)
 					}
@@ -982,7 +1000,7 @@ func runAMQPTopicPublishers(ctx context.Context, params amqpTopicPublishParams) 
 	return counter.Load()
 }
 
-func connectMQTTClient(addr string, protocolVersion byte, clientID string, onMessage func(msg *mqttclient.Message)) (*mqttclient.Client, error) {
+func connectMQTTClient(addr string, protocolVersion byte, clientID string) (*msgclient.Client, error) {
 	if protocolVersion != 4 && protocolVersion != 5 {
 		return nil, fmt.Errorf("unsupported MQTT protocol version %d", protocolVersion)
 	}
@@ -997,52 +1015,48 @@ func connectMQTTClient(addr string, protocolVersion byte, clientID string, onMes
 		SetAutoReconnect(true).
 		SetMessageChanSize(4096)
 
-	if onMessage != nil {
-		opts.SetOnMessageV2(onMessage)
-	}
-
-	c, err := mqttclient.New(opts)
+	c, err := msgclient.NewMQTT(opts)
 	if err != nil {
 		return nil, err
 	}
-	if err := c.Connect(); err != nil {
+	if err := c.Connect(context.Background()); err != nil {
 		return nil, err
 	}
 	return c, nil
 }
 
-func connectAMQPClient(addr string) (*amqpclient.Client, error) {
+func connectAMQPClient(addr string) (*msgclient.Client, error) {
 	opts := amqpclient.NewOptions().
 		SetAddress(addr).
 		SetCredentials("guest", "guest").
 		SetPrefetch(512, 0).
 		SetAutoReconnect(true)
 
-	c, err := amqpclient.New(opts)
+	c, err := msgclient.NewAMQP(opts)
 	if err != nil {
 		return nil, err
 	}
-	if err := c.Connect(); err != nil {
+	if err := c.Connect(context.Background()); err != nil {
 		return nil, err
 	}
 	return c, nil
 }
 
-func disconnectMQTTClients(clients []*mqttclient.Client) {
+func disconnectMQTTClients(clients []*msgclient.Client) {
 	for _, c := range clients {
 		if c == nil {
 			continue
 		}
-		_ = c.Disconnect()
+		_ = c.Close(context.Background())
 	}
 }
 
-func closeAMQPClients(clients []*amqpclient.Client) {
+func closeAMQPClients(clients []*msgclient.Client) {
 	for _, c := range clients {
 		if c == nil {
 			continue
 		}
-		_ = c.Close()
+		_ = c.Close(context.Background())
 	}
 }
